+++ unit-tests/src/test/scala/scala/scalanative/unsafe/ZoneTest.scala
@@ -5,6 +5,7 @@ import org.junit.Test
 import org.junit.Assert._
 
 import scalanative.junit.utils.AssertThrows._
+import scalanative.unsigned._
 
 class ZoneTest {
   private def assertAccessible(bptr: Ptr[_], n: Int) {
@@ -27,7 +28,7 @@ class ZoneTest {
 
   @Test def zoneAllocatorAllocWithApply(): Unit = {
     Zone { implicit z =>
-      val ptr = z.alloc(64 * sizeof[Int])
+      val ptr = z.alloc(64.toUInt * sizeof[Int])
 
       assertAccessible(ptr, 64)
 
@@ -42,7 +43,7 @@ class ZoneTest {
     assertTrue(zone.isOpen)
     assertFalse(zone.isClosed)
 
-    val ptr = zone.alloc(64 * sizeof[Int])
+    val ptr = zone.alloc(64.toUInt * sizeof[Int])
 
     assertAccessible(ptr, 64)
 
@@ -58,10 +59,11 @@ class ZoneTest {
   @Test def allocThrowsExceptionIfZoneAllocatorIsClosed(): Unit = {
     implicit val zone: Zone = Zone.open()
 
-    zone.alloc(64 * sizeof[Int])
+    zone.alloc(64.toUInt * sizeof[Int])
 
     zone.close()
 
-    assertThrows(classOf[IllegalStateException], zone.alloc(64 * sizeof[Int]))
+    assertThrows(classOf[IllegalStateException],
+                 zone.alloc(64.toUInt * sizeof[Int]))
   }
 }
